+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Changed SwiftKeychainWrapper to Valet (Should fix the keychain issue)
 * Fixed bug where timer would continue when user would log out.
 * Added "nothing here" on follower list and following list when a user doesn't have any followers or any one they are following.
+* Added account settings view controller (Doesn't actually contact the api. Just placeholder for now.)
 ## v0.2 (PRERELEASE)
 * Some design changes
 * Added most channel attrbitues that update (except followers and following)